feat: move sealed classes, enums, and data classes to commons (batch 3)
Moved to commons module for KMP iOS migration:

Sealed classes:
- GenericLoadable<T> → commons.ui.components
- UrlPreviewState → commons.ui.components
- ConnectivityStatus → commons.service.connectivity

Enums:
- BookmarkType → commons.ui.bookmarkgroups

Data classes:
- GeohashListCard → commons.model.nip51Lists.geohashLists
- RelayListCard → commons.model.nip51Lists.relayLists

Original files replaced with typealiases for backward compatibility.
All consumers updated with direct imports where needed.
