+++ commons/src/commonMain/kotlin/com/vitorpamplona/amethyst/commons/service/connectivity/ConnectivityStatus.kt
@@ -18,6 +18,15 @@
  * AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.vitorpamplona.amethyst.service.connectivity
+package com.vitorpamplona.amethyst.commons.service.connectivity
 
-typealias ConnectivityStatus = com.vitorpamplona.amethyst.commons.service.connectivity.ConnectivityStatus
+sealed class ConnectivityStatus {
+    data class Active(
+        val networkId: Long,
+        val isMobile: Boolean,
+    ) : ConnectivityStatus()
+
+    object Off : ConnectivityStatus()
+
+    object StartingService : ConnectivityStatus()
+}
